Refactored template functions
