+++ src/view/add-point-button/create-add-point-button-template.js
@@ -2,10 +2,10 @@
  * Creates add new button template
  * @returns {String} template
  */
-const template = () =>
+const createAddPointButtonTemplate = () =>
   `<button
     class="trip-main__event-add-btn btn btn--big btn--yellow"
     type="button">New event
   </button>`;
 
-export {template};
+export {createAddPointButtonTemplate};
