+++ src/view/edit-point/create-destinations-template.js
@@ -1,0 +1,9 @@
+/**
+ * Creates destinations list template
+ * @param {Array<Destination>} destinations - array of available destinations
+ * @returns {String} template
+ */
+const createDestinationsTemplate = (destinations) =>
+  destinations.map(({name}) => `<option value="${name}"></option>`).join('');
+
+export {createDestinationsTemplate};
