+++ src/view/point-list/create-point-list-template.js
@@ -1,0 +1,7 @@
+/**
+ * Creates event list template
+ * @returns {String} template
+ */
+const createPointListTemplate = () => '<ul class="trip-events__list"></ul>';
+
+export {createPointListTemplate};
