--- src/view/point-list/template.js
@@ -1,7 +1,0 @@
-/**
- * Creates event list template
- * @returns {String} template
- */
-const template = () => '<ul class="trip-events__list"></ul>';
-
-export {template};
